requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.10.1
asgiref==3.6.0
billiard==4.2.0
boto3==1.28.73
botocore==1.31.73
celery==5.3.6
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==40.0.2
dataclasses-json==0.5.8
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==4.0.0
Django==4.2
django-allauth==0.52.0
django-ckeditor==6.7.0
django-cors-headers==4.3.0
django-db-logger==0.1.12
django-js-asset==2.1.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
flake8==6.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
jmespath==1.0.1
kombu==5.3.4
Markdown==3.5.2
marshmallow==3.20.1
marshmallow-enum==1.5.1
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==1.6.1
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.0.1
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.12
vine==5.1.0
wcwidth==0.2.12
